refactoring and formatting fixes
