+++ batch/batch/cloud/gcp/worker/disk.py
@@ -1,9 +1,8 @@
 import logging
 from typing import Dict, Optional
 
-from hailtop.utils import check_shell_output, retry_all_errors_n_times
+from hailtop.utils import check_shell_output, LoggingTimer, retry_all_errors_n_times
 from hailtop.aiocloud import aiogoogle
-from hailtop.utils.utils import LoggingTimer
 
 from ....worker.disk import CloudDisk
