+++ batch/batch/driver/instance.py
@@ -11,7 +11,6 @@ from hailtop.utils import time_msecs, time_msecs_str, retry_transient_errors
 from hailtop import httpx
 from gear import Database, transaction
 
-from ..database import check_call_procedure
 from ..globals import INSTANCE_VERSION
 from ..instance_config import InstanceConfig
 from ..cloud.utils import instance_config_from_config_dict
@@ -158,8 +157,8 @@ VALUES (%s, %s);
     async def activate(self, ip_address, timestamp):
         assert self._state == 'pending'
 
-        rv = await check_call_procedure(
-            self.db, 'CALL activate_instance(%s, %s, %s);', (self.name, ip_address, timestamp), 'activate_instance'
+        rv = await self.db.check_call_procedure(
+            'CALL activate_instance(%s, %s, %s);', (self.name, ip_address, timestamp), 'activate_instance'
         )
 
         self.inst_coll.adjust_for_remove_instance(self)
