+++ hail/python/hailtop/utils/__init__.py
@@ -4,7 +4,7 @@ from .utils import (
     bounded_gather, grouped, sync_sleep_and_backoff, sleep_and_backoff, is_transient_error,
     request_retry_transient_errors, request_raise_transient_errors,
     collect_agen, retry_all_errors, retry_transient_errors,
-    retry_long_running, run_if_changed, run_if_changed_idempotent,
+    retry_long_running, run_if_changed, run_if_changed_idempotent, LoggingTimer,
     WaitableSharedPool, RETRY_FUNCTION_SCRIPT, sync_retry_transient_errors,
     retry_response_returning_functions, first_extant_file, secret_alnum_string,
     flatten, filter_none, partition, cost_str, external_requests_client_session, url_basename,
@@ -49,6 +49,7 @@ __all__ = [
     'retry_long_running',
     'run_if_changed',
     'run_if_changed_idempotent',
+    'LoggingTimer',
     'WaitableSharedPool',
     'request_retry_transient_errors',
     'request_raise_transient_errors',
